+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 # Armada Theme Changelog
 
-## [Unreleased]
+[v0.2.0] - 2025-03-17
+
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 - Fleet Dark UI and Editor Color Scheme
 - Plugin Icon
+
+## [Unreleased]
+### Added
+- JavaScript annotator for property reference coloration
+- Colors for Properties including Annotator support for string literals
+
+### Changed